+++ src/commands/cmd_create.js
@@ -3,21 +3,27 @@ const exec = require('child_process');
 const p = require("path");
 
 module.exports = {
+    //* This function creates the aqueduct template files using the Aqueduct command tool 
     create: async function () {
+        //* Asks the user for the name of there project
         let name = await vscode.window.showInputBox({
             prompt: "Enter the name of your project"
         })
 
+        //* Terminate the command if the name is null
         if(name === "") return;
 
+        //* Asks the user to select the location they want to save their new project
         let locationUri = await vscode.window.showOpenDialog({
             canSelectFolders: true,
             canSelectFiles: false,
             canSelectMany: false,
         });
 
+        //* Extracts the file system path from the Uri array
         let path = locationUri[0].fsPath;
 
+        //* Asks the user to select want project type they want for there API
         let template = await vscode.window.showQuickPick(["Default","Database","Database with Authentication"],
         {
             canPickMany: false,
@@ -26,6 +32,7 @@ module.exports = {
         });
 
         let type;
+        //* Assings the type variable to the value of the project type selection. 
         switch (template) {
             case "Default":
                 type = "default";
@@ -40,11 +47,15 @@ module.exports = {
                 type = "default";
                 break;
         }
-
+        
+        //* Checks the current OS.
         if (process.platform === "win32") {
+            //* Checks what drive the selected folder is in
             let drive = `${path[0]}:`;
+            //* If the folder is not in the C drive.
             if (drive !== "c:") {
                 vscode.window.showInformationMessage("We'll take it from here! Creating Aqueduct files. This may take a few minutes.");
+                //* CD into the selected drive, then to the selected path and then run the aqueduct create command.
                 exec.exec(`${drive} && cd ${path} && aqueduct create -t ${type} ${name}`, (err, stdout, stderr) => {
                     if (err) {
                         console.log(err);
@@ -57,6 +68,7 @@ module.exports = {
                     vscode.window.showInformationMessage("Finished! Aqueduct files created successfully");
 
                     let folderUrl = vscode.Uri.file(p.join(path, `${name}`));
+                    //* Opens the created project in VS Code
                     vscode.commands.executeCommand("vscode.openFolder", folderUrl, false)
                 });
             } else {
@@ -76,8 +88,10 @@ module.exports = {
                     vscode.commands.executeCommand("vscode.openFolder", folderUrl, false)
                 });
             }
+        //* Checks if the current OS is either Mac OS or Linux
         } else if (process.platform === "darwin" || process.platform === "linux") {
             vscode.window.showInformationMessage("We'll take it from here! Creating Aqueduct files. This may take take a few minutes.");
+            //* CD into the selected directory and run the aqueduct create command
             exec.exec(`cd ${path} && aqueduct create -t ${type} ${name}`, (err, stdout, stderr) => {
                 if (err) {
                     console.log(err);
@@ -90,11 +104,14 @@ module.exports = {
                 vscode.window.showInformationMessage("Finished! Aqueduct files created successfully");
 
                 let folderUrl = vscode.Uri.file(p.join(path, `${name}`));
+                //* Open the project in VS Code 
                 vscode.commands.executeCommand("vscode.openFolder", folderUrl, false)
             });
         } else {
+            //* Tells the user their OS is not suppoerted and to file a GitHub issue. (Need to know who uses what to see if its viable to add it.)
             let error = await vscode.window.showErrorMessage("Your current OS does not support this feature. Please file a GitHub issue.", "Open GitHub", "Close");
             if (error === "Open GitHub") {
+                //* Opnes GitHub
                 vscode.env.openExternal("https://github.com/AzMoza/aqueduct-helper/issues/new")
             }
         }