+++ CHANGELOG.md
@@ -12,7 +12,7 @@ All notable changes to the Aqueduct Helper will be documented in this file. If y
     - A Quick Pick is now shown when pressed allowing you to select what section of the documentation you want to visit.
 #### Fixed
 - Aqueduct extension showing when not programming in Dart.
-- Pressing escape when creating a model creates undifinded variables.
+- Pressing escape when creating a model creates undefined variables.
 - Snippets being formatted incorrectly.
 - Snippets including typos.